Moved tests in tests/tensor/rewriting
